Fix deduplication in object tracking
We weren't correctly making sure we aren't hooking on_destroy twice due
to a typo and the fact that the sets that keep track of seen functions
were bound to the template scope (they are now global)

Also fixed possible crashes related to _certain_ on destroy calls having
additional arguments.
Due to the way calling convention works on Windowws x86-64, this is safe
to do even if a caller doesn't supply them as the registers involved
would just be unused.